move ScalanAst to the new ast module
