+++ ast/src/main/scala/scalan/meta/Configs.scala
@@ -3,9 +3,9 @@ package scalan.meta
 import java.io.File
 
 import com.trueaccord.lenses.Updatable
-
 import scalan.util.{FileUtil, GraphUtil}
 import scalan.util.StringUtil._
+
 import scala.collection.mutable.{Map => MMap}
 import scalan.meta.ScalanAst.WrapperConf
 
@@ -181,6 +181,7 @@ object UnitConfig {
   }
 }
 
-/** <c>classesPath</c> is a path to JAR or directory containing scalanizer-plugin.properties and a plugin for Scalanizer which implements [scalan.meta.scalanizer.Plugin]. */
+/** <c>classesPath</c> is a path to JAR or directory containing scalanizer-plugin.properties
+  * and a plugin for Scalanizer which implements [scalan.meta.scalanizer.Plugin]. */
 case class PluginConfig(classesPath: String, extraData: Any = null)
