+++ ast/src/main/scala/scalan/meta/LibraryConfig.scala
